Accumulate HPWL in double (host verify + PL kernel)
The host-only check FAILed on superblue11_a (rel_err 2.8e-3) because it
compared two float32-accumulated HPWL sums in different orders; at ~935k
nets summing per-net values (~1e5) into a float is order-dependent to
~0.3%. Both sides were inaccurate -- the packing was fine.

Recompute the verification reference in double (per-net bbox stays float,
positions are float): packed now matches golden to rel_err 0 on both
mgc_pci_bridge32_b and mgc_superblue11_a. Tighten the PASS threshold to
1e-6. Accumulate the PL kernel HPWL in double as well (final metric stored
as float, ~7 sig figs) so hardware matches the reference. Kernel still
C-synthesizes (II=2, Fmax 411 MHz).

+++ vck5000/host/src/pl_algo/src/Packer.cpp
@@ -57,8 +57,8 @@ PackedDesign packDesign(AIEplace::DataBase& db) {
     return pk;
 }
 
-float hpwlFromPacked(const PackedDesign& pk) {
-    float total = 0.0f;
+double hpwlFromPacked(const PackedDesign& pk) {
+    double total = 0.0;
     for (int n = 0; n < pk.header.num_nets; ++n) {
         const int beg = pk.net_ptr[n], end = pk.net_ptr[n + 1];
         if (beg == end) continue;
@@ -72,7 +72,7 @@ float hpwlFromPacked(const PackedDesign& pk) {
             min_x = std::min(min_x, x); max_x = std::max(max_x, x);
             min_y = std::min(min_y, y); max_y = std::max(max_y, y);
         }
-        total += (max_x - min_x) + (max_y - min_y);
+        total += (double)((max_x - min_x) + (max_y - min_y));
     }
     return total;
 }
